Crate sp_clamor

Helper Functions and Types that are used in other Packages of the this Workspace

Modules

clamor

A runtime interface for the Clamor Blockchain

Constants

CID_PREFIX

The IPFS CID prefix used to use to obtain any data that is stored on the Fragnova Blockchain

Functions

get_locked_frag_account

Returns an account ID that can stake FRAG tokens. This returned account ID is determinstically computed from the given account ID (who).

get_vault_id

Get an Account ID deterministically computed from an input hash**.

http_json_post

Make an HTTP POST Request with data body to the URL url

init

Set the Fragnova-owned Geth Node’s URL

Type Definitions

Hash64

64 bytes u8-Array

Hash128

128 bytes u8-Array

Hash256

256 bytes u8-Array

InstanceUnit

Type used to represent an Instance’s Edition ID and an Instance’s Copy ID